fix: address code review findings (C1-C3, I1-I4)
C1: Add core:window:allow-create capability for WebviewWindow
C2: Fix i32/i64 mismatch in ManagedRepo.item_count
C3: Use decorations:true on dashboard window (native close button)
I1: Use i18n for dashboard window title
I2: Simplify doImport to skip open/close validation dance
I3: Remove duplicate addRepo call in openRepo
I4: Add click-outside dismiss for popup menus

+++ src-tauri/src/commands/dashboard.rs
@@ -33,7 +33,7 @@ pub async fn add_managed_repo(
     path: String,
     icon: Option<String>,
     name: Option<String>,
-    item_count: Option<i32>,
+    item_count: Option<i64>,
 ) -> Result<Vec<ManagedRepo>, String> {
     let mut repos = read(&app)?;
     let now = chrono::Utc::now().to_rfc3339();
